\name{HighDimOut-package}
\alias{HighDimOut-package}
\alias{HighDimOut}
\docType{package}
\title{
Outlier Detection Algorithms for High-Dimensional Data
}
\description{
Three high-dimensional outlier detection algorithms and a outlier unification scheme are implemented in this package. The angle-based outlier detection (ABOD) algorithm is based on the work of Kriegel, Schubert, and Zimek [2008]. The subspace outlier detection (SOD) algorithm is based on the work of Kriegel, Kroger, Schubert, and Zimek [2009]. The feature bagging-based outlier detection (FBOD) algorithm is based on the work of Lazarevic and Kumar [2005]. The outlier unification scheme is based on the work of Kriegel, Kroger, Schubert, and Zimek [2011].
}
\details{
\tabular{ll}{
Package: \tab HighDimOut\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-30\cr
License: \tab MIT\cr
}

}
\author{
Cheng Fan

Maintainer: Cheng Fan <raja8885@hotmail.com>

}
\references{
Hans-Peter Kriegel, Matthias Schubert, Arthur Zimek. Angle-based outlier detection in high-dimensional data. KDD 2008, 444-452.

Hans-Peter Kriegel, Peer Kroger, Erich Schubert, Arthur Zimek. Interpreting and Unifying Outlier Scores. SDM 2011, 13-24.

Hans-Peter Kriegel, Peer Kroger, Erich Schubert, Arthur Zimek. Outlier Detection in Axis-Parallel Subspaces of High Dimensional Data. PAKDD 2009, 831-838.

Aleksandar Lazarevic, Vipin Kumar. Feature bagging for outlier detection. KDD 2005, 157-166.
}


