% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighDimOut.R
\docType{data}
\name{TestData}
\alias{TestData}
\title{Testing data for testing the performance of different algorithms}
\format{A data frame with 60 rows and 3 variables:}
\usage{
data(TestData)
}
\description{
The data are generated according to the example in the paper "Kriegel, Kroger, Schubert, and Arthur Zimek, 2009, Outlier Detection in Axis-Parallel Subspaces of High Dimensional Data".
The data has 60 rows and 3 variables. The first two variables are the x and y coordinates.
The third variable indicates the type of data, i.e., "Pattern_1", "Pattern_2", or "Outlier".
The first 25 observations belong to "Pattern_1". Another 25 observations represent "Pattern_2".
The other 10 observations are "Outliers".
}
\examples{
library(ggplot2)
data(TestData)
ggplot(data = TestData, aes(x = x, y = y, shape=Lab, color=Lab)) + geom_point()
}
\keyword{datasets}

